+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Fixed JToggleButtonSelectedProperty value
 * Fixed initial value for ButtonPressedProperty and JTextComponentTextProperty
 * Made the JTextComponentTextProperty fire less events when setting its value
+* Added PropertyValueChangeTrigger to trigger validation when the value of a ReadableProperty changes
 
 # Version 3.2.0